+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/terrains/terrain_importer.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 import numpy as np
 import torch
 import trimesh
-from typing import TYPE_CHECKING, Sequence
+from typing import TYPE_CHECKING
 
 import warp
 from pxr import UsdGeom
@@ -76,6 +76,8 @@ class TerrainImporter:
         self.warp_meshes = dict()
         self.env_origins = None
         self.terrain_origins = None
+        # private variables
+        self._terrain_flat_patches = dict()
 
         # auto-import the terrain based on the config
         if self.cfg.terrain_type == "generator":
@@ -83,10 +85,12 @@ class TerrainImporter:
             if self.cfg.terrain_generator is None:
                 raise ValueError("Input terrain type is 'generator' but no value provided for 'terrain_generator'.")
             # generate the terrain
-            terrain_generator = TerrainGenerator(cfg=self.cfg.terrain_generator)
+            terrain_generator = TerrainGenerator(cfg=self.cfg.terrain_generator, device=self.device)
             self.import_mesh("terrain", terrain_generator.terrain_mesh)
             # configure the terrain origins based on the terrain generator
             self.configure_env_origins(terrain_generator.terrain_origins)
+            # refer to the flat patches
+            self._terrain_flat_patches = terrain_generator.flat_patches
         elif self.cfg.terrain_type == "usd":
             # check if config is provided
             if self.cfg.usd_path is None:
@@ -117,6 +121,17 @@ class TerrainImporter:
         This always returns True.
         """
         return True
+
+    @property
+    def flat_patches(self) -> dict[str, torch.Tensor]:
+        """A dictionary containing the sampled valid (flat) patches for the terrain.
+
+        This is only available if the terrain type is 'generator'. For other terrain types, this feature
+        is not available and the function returns an empty dictionary.
+
+        Please refer to the :attr:`TerrainGenerator.flat_patches` for more information.
+        """
+        return self._terrain_flat_patches
 
     """
     Operations - Visibility.
@@ -159,7 +174,7 @@ class TerrainImporter:
     Operations - Import.
     """
 
-    def import_ground_plane(self, key: str, size: tuple[int, int] = (2.0e6, 2.0e6)):
+    def import_ground_plane(self, key: str, size: tuple[float, float] = (2.0e6, 2.0e6)):
         """Add a plane to the terrain importer.
 
         Args:
@@ -181,7 +196,7 @@ class TerrainImporter:
         self.warp_meshes[key] = convert_to_warp_mesh(mesh.vertices, mesh.faces, device=device)
 
         # get the mesh
-        ground_plane_cfg = sim_utils.GroundPlaneCfg(physics_material=self.cfg.physics_material)
+        ground_plane_cfg = sim_utils.GroundPlaneCfg(physics_material=self.cfg.physics_material, size=size)
         ground_plane_cfg.func(self.cfg.prim_path, ground_plane_cfg)
 
     def import_mesh(self, key: str, mesh: trimesh.Trimesh):
@@ -305,26 +320,6 @@ class TerrainImporter:
         self.env_origins[env_ids] = self.terrain_origins[self.terrain_levels[env_ids], self.terrain_types[env_ids]]
 
     """
-    Operations - Sampling
-    """
-
-    def sample_new_targets(self, env_ids: Sequence[int]) -> torch.Tensor:
-        """Samples terrain-aware locations of flat patches to set spawn or target locations.
-
-        Note:
-            This is a dummy function that returns the environment origins as target locations.
-            Please inherit the class and reimplement the function for specific terrain types
-
-        Args:
-            env_ids: The environment indices to sample targets locations for.
-
-        Returns:
-            The sampled target locations as (x, y, z). Shape is (N, 3).
-        """
-
-        return self.env_origins[env_ids]
-
-    """
     Internal helpers.
     """
 
@@ -356,10 +351,12 @@ class TerrainImporter:
         # create tensor based on number of environments
         env_origins = torch.zeros(num_envs, 3, device=self.device)
         # create a grid of origins
-        num_cols = np.floor(np.sqrt(num_envs))
-        num_rows = np.ceil(num_envs / num_cols)
-        xx, yy = torch.meshgrid(torch.arange(num_rows), torch.arange(num_cols), indexing="xy")
-        env_origins[:, 0] = env_spacing * xx.flatten()[:num_envs] - env_spacing * (num_rows - 1) / 2
-        env_origins[:, 1] = env_spacing * yy.flatten()[:num_envs] - env_spacing * (num_cols - 1) / 2
+        num_rows = np.ceil(num_envs / int(np.sqrt(num_envs)))
+        num_cols = np.ceil(num_envs / num_rows)
+        ii, jj = torch.meshgrid(
+            torch.arange(num_rows, device=self.device), torch.arange(num_cols, device=self.device), indexing="ij"
+        )
+        env_origins[:, 0] = -(ii.flatten()[:num_envs] - (num_rows - 1) / 2) * env_spacing
+        env_origins[:, 1] = (jj.flatten()[:num_envs] - (num_cols - 1) / 2) * env_spacing
         env_origins[:, 2] = 0.0
         return env_origins