+++ source/extensions/omni.isaac.orbit/test/sensors/test_frame_transformer.py
@@ -11,7 +11,7 @@ from __future__ import annotations
 
 """Launch Isaac Sim Simulator first."""
 
-from omni.isaac.orbit.app import AppLauncher
+from omni.isaac.orbit.app import AppLauncher, run_tests
 
 # launch omniverse app
 app_launcher = AppLauncher(headless=True)
@@ -22,10 +22,8 @@ simulation_app = app_launcher.app
 import math
 import scipy.spatial.transform as tf
 import torch
-import traceback
 import unittest
 
-import carb
 import omni.isaac.core.utils.stage as stage_utils
 
 import omni.isaac.orbit.sim as sim_utils
@@ -308,12 +306,4 @@ class TestFrameTransformer(unittest.TestCase):
 
 
 if __name__ == "__main__":
-    try:
-        unittest.main()
-    except Exception as err:
-        carb.log_error(err)
-        carb.log_error(traceback.format_exc())
-        raise
-    finally:
-        # close sim app
-        simulation_app.close()
+    run_tests()