+++ source/extensions/omni.isaac.orbit/test/sensors/test_ray_caster_camera.py
@@ -8,10 +8,10 @@
 
 """Launch Isaac Sim Simulator first."""
 
-from omni.isaac.orbit.app import AppLauncher
+from omni.isaac.orbit.app import AppLauncher, run_tests
 
 # launch omniverse app
-app_launcher = AppLauncher(headless=True)
+app_launcher = AppLauncher(headless=True, offscreen_render=True)
 simulation_app = app_launcher.app
 
 """Rest everything follows."""
@@ -20,10 +20,8 @@ import copy
 import numpy as np
 import os
 import torch
-import traceback
 import unittest
 
-import carb
 import omni.isaac.core.utils.prims as prim_utils
 import omni.isaac.core.utils.stage as stage_utils
 import omni.replicator.core as rep
@@ -372,7 +370,6 @@ class TestWarpCamera(unittest.TestCase):
             spawn=PinholeCameraCfg(
                 focal_length=24.0, focus_distance=400.0, horizontal_aperture=20.955, clipping_range=(1e-4, 1.0e5)
             ),
-            colorize=False,
         )
         camera_usd = Camera(camera_cfg_usd)
 
@@ -442,7 +439,6 @@ class TestWarpCamera(unittest.TestCase):
             spawn=PinholeCameraCfg(
                 focal_length=24.0, focus_distance=400.0, horizontal_aperture=20.955, clipping_range=(1e-6, 1.0e5)
             ),
-            colorize=False,
             offset=CameraCfg.OffsetCfg(pos=(2.5, 2.5, 4.0), rot=offset_rot, convention="ros"),
         )
         camera_usd = Camera(camera_cfg_usd)
@@ -520,7 +516,6 @@ class TestWarpCamera(unittest.TestCase):
             spawn=PinholeCameraCfg(
                 focal_length=24.0, focus_distance=400.0, horizontal_aperture=20.955, clipping_range=(1e-6, 1.0e5)
             ),
-            colorize=False,
             offset=CameraCfg.OffsetCfg(pos=(0, 0, 2.0), rot=offset_rot, convention="ros"),
         )
         prim_usd = prim_utils.create_prim("/World/Camera_usd", "Xform")
@@ -567,12 +562,4 @@ class TestWarpCamera(unittest.TestCase):
 
 
 if __name__ == "__main__":
-    try:
-        unittest.main()
-    except Exception as err:
-        carb.log_error(err)
-        carb.log_error(traceback.format_exc())
-        raise
-    finally:
-        # close sim app
-        simulation_app.close()
+    run_tests()