+++ source/extensions/omni.isaac.orbit/test/sim/test_simulation_context.py
@@ -7,19 +7,17 @@ from __future__ import annotations
 
 """Launch Isaac Sim Simulator first."""
 
-from omni.isaac.orbit.app import AppLauncher
+from omni.isaac.orbit.app import AppLauncher, run_tests
 
 # launch omniverse app
-simulation_app = AppLauncher(headless=True).app
+simulation_app = AppLauncher(headless=True, experience="omni.isaac.sim.python.gym.headless.kit").app
 
 """Rest everything follows."""
 
 import ctypes
 import numpy as np
-import traceback
 import unittest
 
-import carb
 import omni.isaac.core.utils.prims as prim_utils
 from omni.isaac.core.simulation_context import SimulationContext as IsaacSimulationContext
 
@@ -78,7 +76,7 @@ class TestSimulationContext(unittest.TestCase):
         sim = SimulationContext()
         version = sim.get_version()
         self.assertTrue(len(version) > 0)
-        self.assertTrue(version[0] >= 2022)
+        self.assertTrue(version[0] >= 2023)
 
     def test_carb_setting(self):
         """Test setting carb settings."""
@@ -125,14 +123,16 @@ class TestSimulationContext(unittest.TestCase):
         sim.clear_instance()
         self.assertEqual(ctypes.c_long.from_address(id(sim)).value, sim_ref_count - 1)
 
+    def test_zero_gravity(self):
+        """Test that gravity can be properly disabled."""
+        cfg = SimulationCfg(gravity=(0.0, 0.0, 0.0))
+
+        sim = SimulationContext(cfg)
+
+        gravity_dir, gravity_mag = sim.get_physics_context().get_gravity()
+        gravity = np.array(gravity_dir) * gravity_mag
+        np.testing.assert_almost_equal(gravity, cfg.gravity)
+
 
 if __name__ == "__main__":
-    try:
-        unittest.main()
-    except Exception as err:
-        carb.log_error(err)
-        carb.log_error(traceback.format_exc())
-        raise
-    finally:
-        # close sim app
-        simulation_app.close()
+    run_tests()