+++ source/extensions/omni.isaac.orbit/test/utils/test_dict.py
@@ -5,6 +5,18 @@
 
 from __future__ import annotations
 
+# NOTE: While we don't actually use the simulation app in this test, we still need to launch it
+#       because warp is only available in the context of a running simulation
+"""Launch Isaac Sim Simulator first."""
+
+from omni.isaac.orbit.app import AppLauncher, run_tests
+
+# launch omniverse app
+app_launcher = AppLauncher(headless=True)
+simulation_app = app_launcher.app
+
+"""Rest everything follows."""
+
 import unittest
 
 import omni.isaac.orbit.utils.dict as dict_utils
@@ -71,4 +83,4 @@ class TestDictUtilities(unittest.TestCase):
 
 
 if __name__ == "__main__":
-    unittest.main()
+    run_tests()