+++ source/extensions/omni.isaac.orbit/test/utils/test_timer.py
@@ -5,6 +5,18 @@
 
 from __future__ import annotations
 
+# NOTE: While we don't actually use the simulation app in this test, we still need to launch it
+#       because warp is only available in the context of a running simulation
+"""Launch Isaac Sim Simulator first."""
+
+from omni.isaac.orbit.app import AppLauncher, run_tests
+
+# launch omniverse app
+app_launcher = AppLauncher(headless=True)
+simulation_app = app_launcher.app
+
+"""Rest everything follows."""
+
 import time
 import unittest
 
@@ -37,4 +49,4 @@ class TestTimer(unittest.TestCase):
 
 
 if __name__ == "__main__":
-    unittest.main()
+    run_tests()